regression, fehler: cannot cast long to double
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@11615 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: 98778ae9cc56beac9dd5617a07dc4f971051534d

+++ mining/statistics/src/de/uniol/inf/is/odysseus/statistics/physicaloperator/RegressionPO.java
@@ -66,8 +66,9 @@ public class RegressionPO<M extends ITimeInterval, R extends Tuple<M>> extends A
 
 	@Override
 	protected synchronized void process_next(R object, int port) {
-		double x = object.getAttribute(xAttribute);
-		double y = object.getAttribute(yAttribute);
+		 
+		double x = ((Long)object.getAttribute(xAttribute)).doubleValue();
+		double y = ((Long)object.getAttribute(yAttribute)).doubleValue();
 
 		sweepArea.insert(object);
